requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
requests==2.31.0
Werkzeug==3.0.3
python-dotenv==1.0.1
Pillow==10.3.0
